+++ src/AwesomeApp.UI/src/backend/libs/index.ts
@@ -1,3 +1,3 @@
 export * from './IronSessionConfig'
 export * from './ApiMiddlewareHandlers'
-export * from './AwesomeApiService' 
+export * from './AwesomeApiClient'
