+++ src/AwesomeApp.UI/src/shared/libs/AxiosClientHelper.ts
@@ -1,0 +1,7 @@
+import { HttpStatusCode, AxiosRequestConfig } from "axios";
+
+export function acceptStatusCodes(statuses: HttpStatusCode[]): AxiosRequestConfig {
+  return {
+    validateStatus: (status) => statuses.includes(status)   
+  }
+}
