+++ src/AwesomeApp.UI/src/shared/libs/index.ts
@@ -1,1 +1,2 @@
+export * from './AxiosClientHelper'
 export * from './EnvironmentHelpers'
